refactor(config): type auth and translation config reads
Use typed config getters for auth and translation reads where the expected type is fixed, including guard and password broker defaults, auth timebox settings, require-password timeout, locale, and fallback locale.

Cast AUTH_PASSWORD_TIMEOUT in the source auth config so env-provided numeric strings continue to load as integers when consumed through typed config access.

Preserve the nullable default user provider path by keeping auth.defaults.provider on a generic get() read and widening getDefaultUserProvider() to ?string. Add regression coverage for the no-provider path and for AUTH_PASSWORD_TIMEOUT env casting.

+++ src/foundation/config/auth.php
@@ -162,5 +162,5 @@ return [
     |
     */
 
-    'password_timeout' => env('AUTH_PASSWORD_TIMEOUT', 10800),
+    'password_timeout' => (int) env('AUTH_PASSWORD_TIMEOUT', 10800),
 ];
